backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-auth-oauthlib
google-api-python-client
beautifulsoup4
python-dotenv
pydantic
email-validator
requests
